+++ tests/consumers/test_strict_consumer.py
@@ -4,10 +4,8 @@ from confluent_kafka import KafkaError
 from unittest.mock import patch
 from unittest.mock import MagicMock
 
-
-from base import FakeConfluentKafkaConsumer, message
-
 from snuba.consumers.strict_consumer import CommitDecision, NoPartitionAssigned, StrictConsumer
+from tests.backends.confluent_kafka import FakeConfluentKafkaConsumer, build_confluent_kafka_message
 
 
 class TestStrictConsumer:
@@ -28,7 +26,7 @@ class TestStrictConsumer:
     def test_empty_topic(self, create_consumer) -> None:
         kafka_consumer = FakeConfluentKafkaConsumer()
         kafka_consumer.items = [
-            message(0, 0, None, True),
+            build_confluent_kafka_message(0, 0, None, True),
         ]
         create_consumer.return_value = kafka_consumer
 
@@ -56,10 +54,10 @@ class TestStrictConsumer:
         kafka_consumer = FakeConfluentKafkaConsumer()
         create_consumer.return_value = kafka_consumer
 
-        msg = message(0, 0, "ABCABC", False)
+        msg = build_confluent_kafka_message(0, 0, "ABCABC", False)
         kafka_consumer.items = [
             msg,
-            message(0, 0, None, True),
+            build_confluent_kafka_message(0, 0, None, True),
         ]
 
         on_message = MagicMock()
@@ -77,10 +75,10 @@ class TestStrictConsumer:
         error = MagicMock()
         error.code.return_value = KafkaError._PARTITION_EOF
         kafka_consumer.items = [
-            message(0, 0, "ABCABC", False),
-            message(1, 0, "ABCABC", False),
-            message(2, 0, "ABCABC", False),
-            message(0, 0, None, True),
+            build_confluent_kafka_message(0, 0, "ABCABC", False),
+            build_confluent_kafka_message(1, 0, "ABCABC", False),
+            build_confluent_kafka_message(2, 0, "ABCABC", False),
+            build_confluent_kafka_message(0, 0, None, True),
         ]
 
         on_message = MagicMock()
